Turn SearchForm into a function component
