+++ www/src/components/search-form/algolia-styles.js
@@ -1,16 +1,7 @@
 /** @jsx jsx */
-import { jsx } from "theme-ui"
-import { Component } from "react"
-import { navigate } from "gatsby"
 
-import { space } from "../gatsby-plugin-theme-ui"
-import SearchIcon from "./search-icon"
-import { themedInput } from "../utils/styles"
-
-const loadJS = () => import(`./docsearch.min.js`)
-let loadedJs = false
-
-import { Global, css } from "@emotion/core"
+import { space } from "../../gatsby-plugin-theme-ui"
+import { css } from "@emotion/core"
 
 // Override default search result styles (docsearch.css)
 const searchDropdownOffsetTop = space[9]
@@ -292,152 +283,4 @@ const algoliaStyles = t => css`
   }
 `
 
-class SearchForm extends Component {
-  constructor() {
-    super()
-    this.state = { focussed: false }
-    this.autocompleteSelected = this.autocompleteSelected.bind(this)
-  }
-  /**
-   * Replace the default selection event, allowing us to do client-side
-   * navigation thus avoiding a full page refresh.
-   *
-   * Ref: https://github.com/algolia/autocomplete.js#events
-   */ autocompleteSelected(e) {
-    e.stopPropagation()
-    // Use an anchor tag to parse the absolute url (from autocomplete.js) into a relative url
-    const a = document.createElement(`a`)
-    a.href = e._args[0].url
-    this.searchInput.blur()
-    // Compare hash and slug and remove hash if both are same
-    const paths = a.pathname.split(`/`).filter(el => el !== ``)
-    const slug = paths[paths.length - 1]
-    const path =
-      `#${slug}` === a.hash ? `${a.pathname}` : `${a.pathname}${a.hash}`
-    navigate(path)
-  }
-  init() {
-    if (this.algoliaInitialized) {
-      return
-    }
-
-    window.addEventListener(
-      `autocomplete:selected`,
-      this.autocompleteSelected,
-      true
-    )
-    window.docsearch({
-      apiKey: `71af1f9c4bd947f0252e17051df13f9c`,
-      indexName: `gatsbyjs`,
-      inputSelector: `#doc-search`,
-      debug: false,
-      autocompleteOptions: {
-        openOnFocus: true,
-        autoselect: true,
-        hint: false,
-        keyboardShortcuts: [`s`],
-      },
-    })
-    this.algoliaInitialized = true
-  }
-  componentDidMount() {
-    if (
-      typeof window === `undefined` ||
-      typeof window.docsearch === `undefined`
-    ) {
-      // Algolia's docsearch lib not loaded yet so load it.
-      // Lazy load css
-      const path = `https://cdn.jsdelivr.net/npm/docsearch.js@2/dist/cdn/docsearch.min.css`
-      const link = document.createElement(`link`)
-      link.setAttribute(`rel`, `stylesheet`)
-      link.setAttribute(`type`, `text/css`)
-      link.setAttribute(`href`, path)
-      document.head.appendChild(link)
-    }
-  }
-  componentWillUnmount() {
-    window.removeEventListener(
-      `autocomplete:selected`,
-      this.autocompleteSelected,
-      true
-    )
-  }
-  loadAlgoliaJS() {
-    if (!loadedJs) {
-      loadJS().then(a => {
-        loadedJs = true
-        window.docsearch = a.default
-        this.init()
-      })
-    } else {
-      this.init()
-    }
-  }
-  render() {
-    const { focussed } = this.state
-    return (
-      <form
-        sx={{
-          alignItems: `flex-end`,
-          justifyContent: `flex-end`,
-          display: `flex`,
-          flex: [`1 1 auto`, null, `1 0 auto`, null, `0 0 auto`, `1 0 auto`],
-          flexDirection: `row`,
-          mb: 0,
-          mx: [3, null, null, 4],
-          position: `relative`,
-          // minWidth: [null, null, null, null, `12rem`],
-          // maxWidth: [`100%`, `100%`, `100%`, null, `24rem`],
-          "& .algolia-autocomplete": {
-            width: `100%`,
-            display: `block !important`,
-          },
-        }}
-        className="searchWrap"
-        onMouseOver={() => this.loadAlgoliaJS()}
-        onClick={() => this.loadAlgoliaJS()}
-        onSubmit={e => e.preventDefault()}
-      >
-        <Global styles={algoliaStyles} />
-        <label
-          sx={{
-            position: `relative`,
-            width: [`100%`, `100%`, `100%`, focussed ? `14rem` : 24, `100%`],
-            transition: t =>
-              `width ${t.transition.speed.default} ${t.transition.curve.default}, padding ${t.transition.speed.default} ${t.transition.curve.default}`,
-          }}
-        >
-          <input
-            id="doc-search"
-            sx={{
-              ...themedInput,
-              bg: [
-                `themedInput.background`,
-                null,
-                null,
-                focussed ? `themedInput.background` : `transparent`,
-                `themedInput.background`,
-              ],
-              pl: [7, null, null, focussed ? 7 : 24, 7],
-              width: [`100%`, null, null, focussed ? `14rem` : 24, `100%`],
-              transition: t =>
-                `width ${t.transition.speed.default} ${t.transition.curve.default}, padding ${t.transition.speed.default} ${t.transition.curve.default}`,
-            }}
-            type="search"
-            placeholder={`Search gatsbyjs.org`}
-            aria-label="Search gatsbyjs.org"
-            title="Hit 's' to search docs"
-            onFocus={() => this.setState({ focussed: true })}
-            onBlur={() => this.setState({ focussed: false })}
-            ref={input => {
-              this.searchInput = input
-            }}
-          />
-          <SearchIcon focussed={focussed} />
-        </label>
-      </form>
-    )
-  }
-}
-
-export default SearchForm
+export default algoliaStyles
