+++ www/src/components/search-form/index.js
@@ -1,0 +1,1 @@
+export { default } from "./search-form"
